provided functionality to identify provided cluster for ha-tests
